requirements.txt
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
hydra-core>=1.3.0
mlflow>=2.3.0
timm>=0.9.0
scikit-learn
python-dotenv
torchmetrics